init.sh: Validate $project_dir
Fixes #1.

+++ init.sh
@@ -41,6 +41,39 @@ fi
 
 project_name="$1"
 project_dir="$2"
+parent_dir="$(dirname "$project_dir")"
+
+for dir in "$project_dir" "$parent_dir"; do
+  if [[ -d "$dir" ]]; then
+    dir_full="$(cd "$dir" && pwd)"
+    if [[ "$dir_full" == "$script_dir" || "$dir_full" == "$script_dir/"* ]]; then
+      echo "The directory '$script_dir' is for samples only. Can't create a new project here." >&2
+      exit 1
+    fi
+  fi
+done
+
+if [[ ! -d "$parent_dir" ]]; then
+  echo "Can't create a new project in '$project_dir' because the directory '$parent_dir' does not exist." >&2
+  exit 1
+fi
+
+if in_list "$samples" "$project_dir"; then
+  read -p "Are you sure you want to create a project from sample '$sample' in the directory '$project_dir'? [yN] " -n 1 -r >&2
+  echo >&2
+  if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    exit 1
+  fi
+fi
+
+if [[ -d "$project_dir" ]]; then
+  read -p "The directory '$project_dir' already exists! Are you sure you want to create a new pipeline project there? [yN] " -n 1 -r >&2
+  echo >&2
+  if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    exit 1
+  fi
+fi
+
 mkdir -p "$project_dir"
 echo "$project_name" > "$project_dir/.project-name"
